Compartmentalize current date token logic
