+++ src/components/Token/Shared/CurrentDateToken.test.tsx
@@ -3,9 +3,7 @@ import userEvent from '@testing-library/user-event';
 import arrayMutators from 'final-form-arrays';
 import React from 'react';
 import { Form } from 'react-final-form';
-import withIntlConfiguration, {
-  withIntlConfigurationAnyTimezone,
-} from '../../../test/util/withIntlConfiguration';
+import withIntlConfiguration from '../../../test/util/withIntlConfiguration';
 import { HeaderFooterTokenType } from '../../../types/TokenTypes';
 import HeaderFooterCardBody from '../HeaderFooter/HeaderFooterCardBody';
 
@@ -52,42 +50,6 @@ describe('Current date token', () => {
         type: HeaderFooterTokenType.CURRENT_DATE,
         format: 'YEAR_LONG',
         timezone: 'America/Chicago',
-      },
-    });
-  });
-
-  it('Assumes UTC when no known TZ', async () => {
-    const submitter = jest.fn();
-
-    render(
-      withIntlConfigurationAnyTimezone(
-        <Form
-          mutators={{ ...arrayMutators }}
-          onSubmit={(v) => submitter(v)}
-          initialValues={{ test: { type: HeaderFooterTokenType.CURRENT_DATE } }}
-        >
-          {({ handleSubmit }) => (
-            <form onSubmit={handleSubmit}>
-              <HeaderFooterCardBody name="test" />
-              <button type="submit">Submit</button>
-            </form>
-          )}
-        </Form>,
-        'en-US'
-      )
-    );
-
-    expect(
-      screen.getByRole('combobox', { name: 'Timezone' })
-    ).toHaveDisplayValue('UTC');
-
-    await userEvent.click(screen.getByRole('button', { name: 'Submit' }));
-
-    expect(submitter).toHaveBeenLastCalledWith({
-      test: {
-        type: HeaderFooterTokenType.CURRENT_DATE,
-        format: 'YEAR_LONG',
-        timezone: 'UTC',
       },
     });
   });
